Update nPose Core.lslp
Add a fix for Gender where the core will update gender of sitters in seats after assignSlots function does what it needs to do and also after seat changes.

Fixed a bug where prop group 0 (normal props) did not die after a pose change.  This is not exactly as it used to be in V3.10 but is close.  Now when nPose processes the XANIM key word, all group 0 props will die.