Update nPose Adjuster.lslp
Bug Fix.
When the current pose card XANIM line is changed and saved, the adjuster moves but the seated Avatar does not follow this change.  This leads to nasty issues as adjusters and AV are not in sync. 

The fix included is to cause the seated AV to follow the adjuster in this case.