Update nPose NC Reader.lslp
Implement LinkSetData storage.
We only need one item in the CacheNcNames and CacheContent lists as the LinkSetData feature creates our storage.
Two cases where a card must be read: 1) if not found in the LinkSetData storage. 2) Only if the card being read does not match what's found in LinkSetData storage.
Everything else uses what is stored in the LinkSetData storage as a key/value pair.  ncName is used as the key and all contents of the card is the value in the pair.
https://wiki.secondlife.com/wiki/LlLinksetDataListKeys     link to the wiki with the other related links also found on this page.